devtools/client/shared/components/test/mochitest/test_notification_box_03.html

<!DOCTYPE HTML>
<html>
<!--
Test for Notification Box. The test is checking:
* Using custom buttons in a notification
-->
<head>
  <meta charset="utf-8">
  <title>Notification Box</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    let ReactDOM = browserRequire("devtools/client/shared/vendor/react-dom");
    let React = browserRequire("devtools/client/shared/vendor/react");
    let { NotificationBox, PriorityLevels } = browserRequire("devtools/client/shared/components/notification-box");

    // Test rendering
    let boxElement = React.createElement(NotificationBox);
    let notificationBox = TestUtils.renderIntoDocument(boxElement);
    let notificationNode = ReactDOM.findDOMNode(notificationBox);

    let buttonCallbackExecuted = false;
    var buttons = [{
      label: "Button1",
      callback: () => {
        buttonCallbackExecuted = true;

        // Do not close the notification
        return true;
      },
    }, {
      label: "Button2",
      callback: () => {
        // Close the notification (return value undefined)
      },
    }];

    // Append a notification with buttons.
    notificationBox.appendNotification(
      "Info message",
      "id1",
      null,
      PriorityLevels.PRIORITY_INFO_LOW,
      buttons
    );

    let buttonNodes = notificationNode.querySelectorAll(
      ".notification-button");

    is(buttonNodes.length, 2, "There must be two buttons");

    // Click the first button
    TestUtils.Simulate.click(buttonNodes[0]);
    ok(buttonCallbackExecuted, "Button callback must be executed.");

    is(TestUtils.scryRenderedDOMComponentsWithClass(
      notificationBox, "notification").length, 1,
      "There must be one notification");

    // Click the second button (closing the notification)
    TestUtils.Simulate.click(buttonNodes[1]);

    is(TestUtils.scryRenderedDOMComponentsWithClass(
      notificationBox, "notification").length, 0,
      "The notification box must be empty now");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
